[ET-43] refactor(reservation) : 필드명 수정 및 이벤트 페이로드 반환 포맷 수정
- 불필요한 Wrapper클래스 삭제
- SeatPreemptPayload로 이름 수정
- SeatPreemptPayload UUID를 리스트로반환
